#11929: Move work_split_tilize
