+++ ttnn/cpp/ttnn/operations/core/work_split/work_split_tilize.hpp
@@ -11,7 +11,9 @@
 #include "ttnn/tensor/types.hpp"
 #include "tt_metal/common/core_coord.h"
 
-namespace tt::tt_metal {
+namespace ttnn {
+
+namespace operations::core::work_split {
 
 struct BlockSplit {
     uint32_t ncores;
@@ -174,7 +176,7 @@ struct FullRep {
 };
 
 inline std::vector<std::vector<BlockRep>> distribute_work(
-    const Shape& unpadded, const Padding& padding, uint32_t num_cores, uint32_t blocks_per_core, bool has_cliff, uint32_t nblocks_per_core_cliff) {
+    const tt::tt_metal::Shape& unpadded, const Padding& padding, uint32_t num_cores, uint32_t blocks_per_core, bool has_cliff, uint32_t nblocks_per_core_cliff) {
     auto input_w = unpadded.rank() >= 4 ? unpadded[-4] : 1;
     auto input_z = unpadded.rank() >= 3 ? unpadded[-3] : 1;
     auto input_y = unpadded.rank() >= 2 ? unpadded[-2] : 1;
@@ -228,4 +230,8 @@ inline std::vector<std::vector<BlockRep>> distribute_work(
     return core_assignments;
 }
 
-}  // namespace tt::tt_metal
+} // namespace operations::core::work_split
+
+using namespace operations::core::work_split;
+
+}  // namespace ttnn
